Gen-AI-Projects/Quiz-App-Pinecone/requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai
langchain
langchain-openai
langchain-pinecone
langchain-community
pandas 
numpy
python-dotenv
streamlit 
langchain-text-splitters 
 

